% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_frustration}
\alias{calculate_frustration}
\title{Calculate Frustration}
\usage{
calculate_frustration(
  PdbFile = NULL,
  PdbID = NULL,
  Chain = NULL,
  Electrostatics_K = NULL,
  seqdist = 12,
  Modes = "configurational",
  Graphics = TRUE,
  ResultsDir
)
}
\arguments{
\item{PdbFile}{File containing the protein structure. The full path to the file is needed}

\item{PdbID}{PdbID of the protein structure.}

\item{Chain}{Chain of the protein structure.}

\item{Electrostatics_K}{K constant to use in the electrostatics mode. Default: NULL (no electrostatics is considered).}

\item{seqdist}{Sequence at which contacts are considered to interact (3 or 12). Default: 12}

\item{Modes}{Local frustration index to be calculated (configurational, mutational, singleresidue). Default=configurational}

\item{Graphics}{The corresponding graphics are made (TRUE or FALSE). Default: TRUE}

\item{ResultsDir}{Path to the folder where results will be stored.}
}
\value{
Pdb Frustration object
}
\description{
Calculates local energetic frustration for a protein structure
}
