% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_backbone_complete}
\alias{check_backbone_complete}
\title{Check Backbone Complete.}
\usage{
check_backbone_complete(Pdb)
}
\arguments{
\item{Pdb}{Pdb frustration object.}
}
\value{
Flag indicating if the backbone should be completed.
}
\description{
Checks the backbone of a given protein structure to be processed by the frustratometeR pipeline.
}
