% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{complete_backbone}
\alias{complete_backbone}
\title{Complete Backbone}
\usage{
complete_backbone(Pdb)
}
\arguments{
\item{Pdb}{Frustration object}
}
\value{
Pdb Frustration object with backbone completed and flag indicated if the backbone was completed
}
\description{
Completes the backbone of a given protein structure to be processed by the frustratometer pipeline
}
