% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{detect_dynamic_clusters}
\alias{detect_dynamic_clusters}
\title{Detect Dynamic clusters}
\usage{
detect_dynamic_clusters(
  Dynamic = Dynamic,
  LoessSpan = 0.05,
  FiltDifprob = 0.7,
  FiltMean = 0.5,
  Ncp = 10,
  CorThreshold = 0.8,
  LeidenResol = 2
)
}
\arguments{
\item{LeidenResol}{}
}
\value{
Dynamic Frustration Object
}
\description{

}
