% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dir_frustration}
\alias{dir_frustration}
\title{Calculate Directory Frustration.}
\usage{
dir_frustration(
  PdbsDir,
  OrderList = NULL,
  Chain = NULL,
  Electrostatics_K = NULL,
  SeqDist = 12,
  Mode = "configurational",
  Graphics = TRUE,
  Visualization = TRUE,
  ResultsDir
)
}
\arguments{
\item{PdbsDir}{Directory containing all protein structures. The full path to the file is needed.}

\item{OrderList}{Orderer list of PDB files to calculate frustration. If it is NULL, frustration is calculated for all PDBs. Type: vector. Default: NULL.}

\item{Chain}{Chain or Chains of the protein structure. Type: character/vector of characters. Default: NULL.}

\item{Electrostatics_K}{K constant to use in the electrostatics Mode. Type: numeric. Default: NULL (no electrostatics is considered).}

\item{SeqDist}{Sequence at which contacts are considered to interact (3 or 12). Type: numeric. Default: 12.}

\item{Mode}{Local frustration index to be calculated (configurational, mutational, singleresidue). Default: configurational.}

\item{Graphics}{The corresponding graphics are made (TRUE or FALSE). Type: bool. Default: TRUE.}

\item{Visualization}{Make visualizations, including pymol. Type: bool. Default: TRUE.}

\item{ResultsDir}{Path to the folder where results will be stored.}
}
\value{
Files resulting from the frustration of each structure in ResultsDir.
}
\description{
Calculate local energy frustration for all protein structures in one directory.
}
