% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dynamic_frustration}
\alias{dynamic_frustration}
\title{Dynamic Frustration}
\usage{
dynamic_frustration(
  PdbDir = NULL,
  OrderList = NULL,
  Electrostatics_K = NULL,
  seqdist = 12,
  Modes = "configurational",
  Resno = NULL,
  Chain = NULL,
  GIFs = FALSE,
  ResultsDir = NULL
)
}
\arguments{
\item{PdbDir}{File containing all proteins structures. The full path to the file is needed}

\item{OrderList}{Orderer list of PDB files to calculate frustration. If it is NULL, frustration is calculated for all PDBs. Default: NULL}

\item{Electrostatics_K}{K constant to use in the electrostatics mode. Default: NULL (no electrostatics is considered).}

\item{seqdist}{Sequence at which contacts are considered to interact (3 or 12). Default: 12}

\item{Modes}{Local frustration index to be calculated (configurational, mutational, singleresidue). Default: configurational}

\item{Resno}{Resno of residue to analyze. If it is NULL, not analyze. Default: NULL}

\item{Chain}{Chain of residue to analyze}

\item{GIFs}{if it is TRUE, gifs of contact maps and proportion 5 adens are made for all the frames. Default: FALSE}

\item{ResultsDir}{Path to the folder where results will be stored.}
}
\description{
Calculates local energetic frustration for a dynamic
}
