% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dynamic_res}
\alias{dynamic_res}
\title{Dynamic Res.}
\usage{
dynamic_res(Dynamic, Resno, Chain, Graphics = TRUE)
}
\arguments{
\item{Dynamic}{Dynamic frustration object.}

\item{Resno}{Residue specific analyzed. Type: numeric.}

\item{Chain}{Chain of specific residue. Type: character.}

\item{Graphics}{If it is TRUE, the graphs corresponding to the residual frustration in the dynamics are made and stored according to the frustration index used. Type: bool. Default: TRUE.}
}
\value{
Dynamic frustration object adding Mutations attribute for the residue of the indicated chain.
}
\description{
Obtain the local frustration of a specific residue in the dynamics.
}
