% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{frustra_movie}
\alias{frustra_movie}
\title{Frustration movie in Pymol.}
\usage{
frustra_movie(Dynamic)
}
\arguments{
\item{Dynamic}{Dynamic frustration object.}
}
\value{
Stores pymol scripts representations_(Mode).pml and GenerateMovie_(Mode).pml
}
\description{
It generates visualization in pymol of the frames of the indicated dynamics and script for its execution in pymol of generation of the film.
}
