% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_clusters}
\alias{get_clusters}
\title{Get clusters information}
\usage{
get_clusters(Dynamic, Clusters = "all")
}
\arguments{
\item{Dynamic}{Dynamic Frustration Object.}

\item{Clusters}{Indicates the clusters, for example, c(1, 2, 3), clusters 1, 2 and 3. Default: "all".}
}
\value{
Data frame containing name, number and cluster of each residue belonging to Clusters
}
\description{
Obtain information about the clusters obtained from detect_dynamic_clusters(), 
the name and number of the residues belonging to each cluster indicated in Clusters.
}
