% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_frustration}
\alias{get_frustration}
\title{Get frustration data}
\usage{
get_frustration(Pdb = NULL)
}
\arguments{
\item{Pdb}{}
}
\value{
Frustration table
}
\description{
Returns a data frame corresponding to the frustration data previously calculated by calculate_frustration
}
