% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_frustration_dynamic}
\alias{get_frustration_dynamic}
\title{Get frustration dynamic.}
\usage{
get_frustration_dynamic(Dynamic, Resno, Chain, Frames = NULL)
}
\arguments{
\item{Dynamic}{Dynamic frustration object obtained by dynamic_frustration().}

\item{Resno}{Specific residue. Type: numeric.}

\item{Chain}{Specific chain. Type: character.}

\item{Frames}{Specific frames. Type: vector.}
}
\value{
Frustration table. Type: data frame.
}
\description{
Obtains and returns the table of frustration of a specific chain and residue in a complete dynamic or in the indicated frames.
The frustration of a specific residue must have previously been calculated using dynamic_res().
}
