% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{gif_5adens_proportions}
\alias{gif_5adens_proportions}
\title{5Adens proportions gif}
\usage{
gif_5adens_proportions(
  PdbDir = NULL,
  OrderList = NULL,
  Modes = "configurational"
)
}
\arguments{
\item{PdbDir}{File containing all proteins structures. The full path to the file is needed}

\item{OrderList}{Orderer list of PDB files to calculate frustration.}

\item{Modes}{Local frustration index to be calculated (configurational, mutational, singleresidue). Default: configurational}
}
\description{
Generate Gif of 5Adens proportions plot of the structure under study along the dynamics
}
