% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{gif_contact_map}
\alias{gif_contact_map}
\title{Contact map gif}
\usage{
gif_contact_map(PdbDir = NULL, OrderList = NULL, Modes = "configurational")
}
\arguments{
\item{PdbDir}{File containing all proteins structures. The full path to the file is needed}

\item{OrderList}{Orderer list of PDB files to calculate frustration.}

\item{Modes}{Local frustration index to be calculated (configurational, mutational, singleresidue). Default: configurational}
}
\description{
Generate Gif of contact map of the structure under study along the dynamics
}
