% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pdb_equivalences}
\alias{pdb_equivalences}
\title{Pdb Equivalences}
\usage{
pdb_equivalences(Pdb)
}
\arguments{
\item{Pdb}{Frustration object}
}
\value{
Pdb equivalences
}
\description{
Internal function that produces auxiliar files to run the frustratometer pipeline.
}
