% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_5Adens_proportions}
\alias{plot_5Adens_proportions}
\title{Plot 5Adens proportions}
\usage{
plot_5Adens_proportions(Pdb, Chain = NULL)
}
\arguments{
\item{Pdb}{Frustration object}

\item{Chain}{Chain of residue to analyze. Default: NULL}
}
\description{
Generates plot to analyze the density of contacts around a sphere of 5 Armstrongs, centered in the C-alfa atom from the residue. The different classes of contacts based on the mutational frustration index are counted in relative terms.
}
