% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_5Andens}
\alias{plot_5Andens}
\title{Plot 5Adens.}
\usage{
plot_5Andens(Pdb, Chain = NULL, Show = TRUE)
}
\arguments{
\item{Pdb}{Pdb Frustration object.}

\item{Chain}{Chain of residue. Type: character. Default: NULL.}

\item{Show}{Print graphic on screen(TRUE), but(FALSE) store in corresponding directory. Type: bool. Default: TRUE.}
}
\value{
ggplot2 object.
}
\description{
Generates plot to analyze the density of contacts around a sphere of 5 Armstrongs, centered in the C-alfa atom from the residue.
 The different classes of contacts based on the mutational frustration index are counted in absolute terms.
}
