% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_clusters_pymol}
\alias{plot_clusters_pymol}
\title{Plot clusters in pymol}
\usage{
plot_clusters_pymol(Dynamic, Clusters = "all")
}
\arguments{
\item{Dynamic}{Dynamic Frustration Object.}

\item{Clusters}{Indicates the clusters, for example, c(1, 2, 3), clusters 1, 2 and 3. Default: "all".}
}
\description{
Generates a pymol session to observe the residues belonging to the clusters indicate in Clusters.
}
