% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_delta_frus}
\alias{plot_delta_frus}
\title{Delta frustration of mutated residue}
\usage{
plot_delta_frus(PdbPath = NULL, DataDir = NULL, ResultDir = NULL, Chain = NULL)
}
\arguments{
\item{PdbPath}{Full path of the Pdb to analyze}

\item{DataDir}{Full directory path where the singleresidue tables generated by mutate_res are located for the n residues analyzed}

\item{ResultDir}{Full path of the directory where the resulting graphic will be stored}

\item{Chain}{Chain of residue to analyze. Default: NULL}
}
\description{
Generate a graph of the singleresidue frustration difference for the mutation of the n mutated residues by using mutate_res
}
