% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_dynamic_clusters_graph}
\alias{plot_dynamic_clusters_graph}
\title{Plot Graph Dynamic Clusters}
\usage{
plot_dynamic_clusters_graph(Dynamic = Dynamic, Save = FALSE)
}
\arguments{
\item{Dynamic}{Dynamic Frustration Object}

\item{Save}{If it is TRUE it saves the graph, otherwise it does not. Type: bool. Default: FALSE.}
}
\value{
ggplot2 object.
}
\description{

}
