% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_dynamic_res}
\alias{plot_dynamic_res}
\title{Plot Dynamic Res.}
\usage{
plot_dynamic_res(Dynamic, Resno, Chain, Save = FALSE)
}
\arguments{
\item{Dynamic}{Dynamic frustration object.}

\item{Resno}{Specific residue. Type: numeric.}

\item{Chain}{Specific chain. Type: character.}

\item{Save}{If it is TRUE it saves the graph, otherwise it does not. Type: bool. Default: FALSE.}
}
\value{
ggplot2 object.
}
\description{
Graph the local frustration of a specific residual in the dynamics.
}
