% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_dynamic_res_5Adens_proportion}
\alias{plot_dynamic_res_5Adens_proportion}
\title{Plot 5Adens proportion for specific residue.}
\usage{
plot_dynamic_res_5Adens_proportion(Dynamic, Resno, Chain, Save = FALSE)
}
\arguments{
\item{Dynamic}{Dynamic frustration object.}

\item{Resno}{Specific residue. Type: numeric.}

\item{Chain}{Specific chain. Type: character.}

\item{Save}{If it is TRUE it saves the graph, otherwise it does not. Type: bool. Default: FALSE.}
}
\value{
ggplot2 object.
}
\description{
Generates plot to analyze the density of contacts around a sphere of 5 Armstrongs, centered in the C-alfa atom from the residue in Dynamic.
}
