% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_mutate_res}
\alias{plot_mutate_res}
\title{Frustration of mutated residue}
\usage{
plot_mutate_res(Pdb = NULL, Mutation = NULL, Show = TRUE)
}
\arguments{
\item{Pdb}{Pdb object resulting from calculate_frustration}

\item{Mutation}{Mutation attribute specific to the Pdb object}

\item{Show}{Print plot on screen. Default: TRUE}
}
\description{
Generate a graph of the configurational or mutational frustration difference for the mutation of the n mutated residues by using mutate_res
}
