% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_res_dynamics}
\alias{plot_res_dynamics}
\title{Plot dynamics residues}
\usage{
plot_res_dynamics(Dynamic = Dynamic, Resno, Chain, Save = FALSE)
}
\arguments{
\item{Dynamic}{Dynamic Frustration Object}

\item{Resno}{Specific residue. Type: numeric.}

\item{Chain}{Specific chain. Type: character.}

\item{Save}{If it is TRUE it saves the graph, otherwise it does not. Type: bool. Default: FALSE.}
}
\value{
ggplot2 object.
}
\description{
Plots the singleresidue frustration index for a residual of a specific chain along the dynamics and fitted model in the detect_dynamic_clusters() function.
}
