% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_variable_res_filter}
\alias{plot_variable_res_filter}
\title{Plot variables residues filter}
\usage{
plot_variable_res_filter(Dynamic = Dynamic, Save = FALSE)
}
\arguments{
\item{Dynamic}{Dynamic Frustration Object}

\item{Save}{If it is TRUE it saves the graph, otherwise it does not. Type: bool. Default: FALSE.}
}
\value{
ggplot2 object.
}
\description{
Plot the filter in frustration dynamic range and the mean of the singleresidue index for each residue along the dynamics.
}
