% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{replace_Expr}
\alias{replace_Expr}
\title{Expression replacement in a file}
\usage{
replace_Expr(pattern, replacement, file)
}
\arguments{
\item{pattern}{Pattern string to replace.}

\item{replacement}{The character string to be replaced.}

\item{file}{Full path of the file where the pattern will be replaced.}
}
\value{
File with pattern replaced by replacement.
}
\description{
Search and replace pattern by replacement in the file lines.
}
