% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{save_res_dynamic_clusters}
\alias{save_res_dynamic_clusters}
\title{Save plots of the dynamics of the residues belonging to a cluster}
\usage{
save_res_dynamic_clusters(Dynamic, Clusters = "all", Ncol = 2, Nrow = 2)
}
\arguments{
\item{Dynamic}{Dynamic Frustration Object}

\item{Clusters}{Indicates the clusters, for example, c(1, 2, 3), clusters 1, 2 and 3. Default: "all".}

\item{Ncol}{Number of columns on each page of the file output. Default: 2.}

\item{Nrow}{Number of rows on each page of the file output. Default: 2.}
}
\description{
It graphs the frustration dynamics of each residue belonging to the indicated clusters (Clusters) and stores the results in files in PDF format.
}
