% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{view_frustration_pymol}
\alias{view_frustration_pymol}
\title{View Frustration Pymol}
\usage{
view_frustration_pymol(pdb)
}
\arguments{
\item{Pdb}{Frustration object}
}
\description{
Generates a pymol session to observe the frustration patterns on top of the pdb protein structure
}
