% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{XAdens}
\alias{XAdens}
\title{X Adens.}
\usage{
XAdens(Pdb, ratio = 5)
}
\arguments{
\item{Pdb}{Pdb frustration object.}

\item{ratio}{Sphere radius. Type: numeric.}
}
\value{
File (Mode)_5adens.
}
\description{
Calculate the proportion of each type of contact (neutral, highly minimally frustrated) of each residue Pdb within a sphere of radius ratio
Is performed by obtaining information from the "tertiary_frustration.dat" file (intermediate processing).
}
