% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_frustration}
\alias{calculate_frustration}
\title{Calculate Frustration.}
\usage{
calculate_frustration(
  PdbFile = NULL,
  PdbID = NULL,
  Chain = NULL,
  Electrostatics_K = NULL,
  SeqDist = 12,
  Mode = "configurational",
  Graphics = TRUE,
  Visualization = TRUE,
  ResultsDir = NULL
)
}
\arguments{
\item{PdbFile}{Full path of the file containing the protein structure. Default: NULL.}

\item{PdbID}{PdbID of the protein structure. Default: NULL.}

\item{Chain}{Chain or Chains of the protein structure. Type: character/vector of characters. Default: NULL.}

\item{Electrostatics_K}{K constant to use in the electrostatics Mode. Type: numeric. Default: NULL (no electrostatics is considered).}

\item{SeqDist}{Sequence at which contacts are considered to interact (3 or 12). Type: numeric. Default: 12.}

\item{Mode}{Local frustration index to be calculated (configurational, mutational, singleresidue). Type: string. Default: configurational.}

\item{Graphics}{The corresponding graphics are made (TRUE or FALSE). Type: bool. Default: TRUE.}

\item{Visualization}{Make visualizations, including pymol. Type: bool. Default: TRUE.}

\item{ResultsDir}{Path to the folder where results will be stored. If not specified, it will be stored in the directory returned by tempdir(). Default: NULL.}
}
\value{
Pdb frustration object.
}
\description{
Calculates local energetic frustration for a protein structure.
}
