% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dir_frustration}
\alias{dir_frustration}
\title{Calculate Directory Frustration}
\usage{
dir_frustration(
  PdbDir = NULL,
  OrderList = NULL,
  Electrostatics_K = NULL,
  seqdist = 12,
  Modes = "configurational",
  Graphics = TRUE,
  Visualization = TRUE,
  ResultsDir = NULL
)
}
\arguments{
\item{PdbDir}{File containing all proteins structures. The full path to the file is needed}

\item{OrderList}{Orderer list of PDB files to calculate frustration. If it is NULL, frustration is calculated for all PDBs. Default: NULL}

\item{Electrostatics_K}{K constant to use in the electrostatics mode. Default: NULL (no electrostatics is considered).}

\item{seqdist}{Sequence at which contacts are considered to interact (3 or 12). Default: 12}

\item{Modes}{Local frustration index to be calculated (configurational, mutational, singleresidue). Default: configurational}

\item{Graphics}{The corresponding graphics are made (TRUE or FALSE). Default: TRUE}

\item{Visualization}{Make visualizations, including pymol. Default: TRUE}

\item{ResultsDir}{Path to the folder where results will be stored.}
}
\description{
Calculate local energy frustration for all protein structures in one directory
}
