% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dynamic_frustration}
\alias{dynamic_frustration}
\title{Dynamic Frustration.}
\usage{
dynamic_frustration(
  PdbsDir,
  OrderList = NULL,
  Chain = NULL,
  Electrostatics_K = NULL,
  SeqDist = 12,
  Mode = "configurational",
  GIFs = FALSE,
  ResultsDir
)
}
\arguments{
\item{PdbsDir}{Directory containing all protein structures. The full path to the file is needed.}

\item{OrderList}{Orderer list of PDB files to calculate frustration. If it is NULL, frustration is calculated for all PDBs. Type: vector. Default: NULL.}

\item{Chain}{Chain of the protein structure. Type: character. Default: NULL.}

\item{Electrostatics_K}{K constant to use in the electrostatics Mode. Type: numeric. Default: NULL (no electrostatics is considered).}

\item{SeqDist}{Sequence at which contacts are considered to interact (3 or 12). Type: numeric. Default: 12.}

\item{Mode}{Local frustration index to be calculated (configurational, mutational, singleresidue). Default: configurational.}

\item{GIFs}{If it is TRUE, the contact map gifs and 5 adens proportion of all the frames of the dynamic will be stored, otherwise they will not be stored. Type: bool. Default: FALSE}

\item{ResultsDir}{Path to the folder where results will be stored.}
}
\value{
Dynamic frustration object
}
\description{
Calculates local energetic frustration for a dynamic.
}
