% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{gif_5adens_proportions}
\alias{gif_5adens_proportions}
\title{5Adens proportions gif.}
\usage{
gif_5adens_proportions(Dynamic, Show = FALSE)
}
\arguments{
\item{Dynamic}{Dynamic frustration object.}

\item{Show}{Print Gif on screen(TRUE), but(FALSE) store in corresponding directory. Type: bool. Default: FALSE.}
}
\description{
Generate Gif of 5Adens proportions plot of the structure under study along the dynamics.
}
