% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{gif_contact_map}
\alias{gif_contact_map}
\title{Contact map gif.}
\usage{
gif_contact_map(Dynamic, Show = FALSE)
}
\arguments{
\item{Dynamic}{Dynamic frustration object.}

\item{Show}{Print Gif on screen(TRUE), but(FALSE) store in corresponding directory. Type: bool. Default: FALSE.}
}
\description{
Generate Gif of contact map of the structure under study along the dynamics.
}
