% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mutate_res}
\alias{mutate_res}
\title{Frustration of mutated residue}
\usage{
mutate_res(
  PdbPath = NULL,
  JobDir = NULL,
  Modes = "configurational",
  Chain = NULL,
  Resno = NULL,
  Split = TRUE,
  Method = "Threading"
)
}
\arguments{
\item{PdbPath}{Pdb file path}

\item{JobDir}{Full directory path where the results will be stored}

\item{Modes}{Local frustration index to be calculated (configurational, mutational, singleresidue). Default: configurational}

\item{Chain}{Chain of residue to analyze. Default: NULL}

\item{Resno}{Resno of residue to analyze. If it is NULL, not analyze. Default: NULL}

\item{Split}{Split that you are going to calculate frustration. If it is TRUE specific string, if it is FALSE full complex. Default: TRUE}

\item{Method}{Method indicates the method to use to perform the mutation (Threading or Modeller). Default: Threading}
}
\value{
Table of frustration results in file
}
\description{
Calculates local energetic frustration for residue mutation
}
