% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mutate_res}
\alias{mutate_res}
\title{Frustration of mutated residue}
\usage{
mutate_res(Pdb, Resno, Chain, Split = TRUE, Method = "Threading")
}
\arguments{
\item{Pdb}{Pdb frustration object.}

\item{Resno}{Resno of the residue to be mutated. Type: numeric. Default: NULL.}

\item{Chain}{Chain of the residue to be mutated. Type: character. Default: NULL.}

\item{Split}{Split that you are going to calculate frustration. If it is TRUE specific string, if it is FALSE full complex. Default: TRUE.}

\item{Method}{Method indicates the method to use to perform the mutation (Threading or Modeller). Default: Threading}
}
\value{
Returns Pdb frustration object with corresponding Mutation attribute
}
\description{
Calculate the local energy frustration for each of the 20 residual variants in the Resno position and Chain chain.
Use the frustration index indicated in the Pdb object.
}
