% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pdb_equivalences}
\alias{pdb_equivalences}
\title{Pdb Equivalences.}
\usage{
pdb_equivalences(Pdb)
}
\arguments{
\item{Pdb}{Pdb frustration object.}
}
\value{
Pdb equivalences in matrix and file "pdb_equivalences.txt".
}
\description{
Internal function that produces auxiliary files in the execution of the script, numerical equivalences between the Pdb and its sequence.
}
