% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_dynamic_clusters_graph}
\alias{plot_dynamic_clusters_graph}
\title{Plot Graph Dynamic Clusters}
\usage{
plot_dynamic_clusters_graph(Dynamic)
}
\arguments{
\item{Dynamic}{Dynamic Frustration Object}
}
\description{
Plot the graph obtained through detect_dynamic_clusters(),
 where each color represents a cluster
}
