% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_dynamic_res}
\alias{plot_dynamic_res}
\title{View Dynamic Frustration of Resid}
\usage{
plot_dynamic_res(DynDir = NULL, Resno = NULL, Modes = "configurational")
}
\arguments{
\item{DynDir}{Full path of the table resulting from the execution of dynamic_frustration for the corresponding residue}

\item{Resno}{Resno of residue to analyze.}

\item{Modes}{Local frustration index to be calculated (configurational, mutational, singleresidue). Default: configurational}
}
\description{
Generate graphs corresponding to the frustration of the residue along the dynamics
}
