% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_mutate_res}
\alias{plot_mutate_res}
\title{Frustration of mutated residue}
\usage{
plot_mutate_res(
  Pdb,
  Resno,
  Chain,
  Method = "Threading",
  DeltaFrus = FALSE,
  Show = TRUE
)
}
\arguments{
\item{Pdb}{Pdb frustration object}

\item{Resno}{Specific residue. Type: numeric.}

\item{Chain}{Specific chain. Type: character.}

\item{Method}{Method indicates the method to use to perform the mutation (Threading or Modeller). Default: Threading}

\item{Show}{Print Gif on screen(TRUE), but(FALSE) store in corresponding directory. Type: bool. Default: FALSE.}
}
\value{
ggplot2 object.
}
\description{
Plot the frustration for each of the 20 residue variants at a given position in the structure.
}
