% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_mutate_res}
\alias{plot_mutate_res}
\title{Frustration of mutated residue}
\usage{
plot_mutate_res(
  PdbPath = NULL,
  DataFile = NULL,
  ResultDir = NULL,
  Chain = NULL,
  Modes = "configurational",
  Resno = NULL
)
}
\arguments{
\item{PdbPath}{Full path of the Pdb to analyze}

\item{DataFile}{Full path of the table resulting from the execution of mutate_res to parse}

\item{ResultDir}{Full path of the directory where the resulting graphic will be stored}

\item{Chain}{Chain of residue to analyze. Default: NULL}

\item{Modes}{Local frustration index to be calculated (configurational, mutational). Default: configurational}

\item{Resno}{Resno of residue to analyze.}
}
\description{
Generate a graph of the configurational or mutational frustration difference for the mutation of the n mutated residues by using mutate_res
}
