% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{replace_Expr}
\alias{replace_Expr}
\title{Replace expression}
\usage{
replace_Expr(pattern = NULL, replacement = NULL, file = NULL)
}
\arguments{
\item{pattern}{Pattern string to replace}

\item{replacement}{Character string to be replaced}

\item{file}{Full path of the file where it will be replaced}
}
\value{

}
\description{
Searches for a string in a file and replaces
}
