% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{XAdens}
\alias{XAdens}
\title{Replace expression}
\usage{
XAdens(Pdb = NULL, ratio = 5)
}
\arguments{
\item{Pdb}{Pdb object}

\item{ratio}{Radius of the sphere for calculating proportions}
}
\value{

}
\description{
Calculate the proportion by type of contact in a sphere of radius X
}
